requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
fastapi==0.104.1
h11==0.16.0
idna==3.11
pillow==11.3.0
pydantic==2.12.2
pydantic_core==2.41.4
python-dotenv==1.1.1
python-multipart==0.0.6
numpy>=1.26.0
opencv-python>=4.8.0
decorator>=4.4.2
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
proglog>=0.1.10
tqdm>=4.66.0
